Fixed: move flags to /data
The file flags were stored in the container directly: if you mount the
database directory from your host, the file flags are only stored in the
container.

The problem with this approach is that the container can be out-of-sync
with the actual database state. For instance, if you re-create the
container, the script may load the data in the toLoad directory
despite they have been loaded already because the flags have been
removed with the containers.

The inverse problem may occurs too: if you stop the container, delete
the database directory and restart the container, then the data in
toLoad directory will not be loaded because it considers they have been
loaded already because the flags are still there.